add SwapSummary component
