+++ src/hooks/useOperationModal/Modal/RepayForm/SubmitSection/index.tsx
@@ -1,14 +1,12 @@
 /** @jsxImportSource @emotion/react */
-import { Typography } from '@mui/material';
 import { ApproveTokenSteps, PrimaryButton } from 'components';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'translation';
 import { Swap, Token } from 'types';
-import { areTokensEqual, convertWeiToTokens, getSwapRouterContractAddress } from 'utilities';
+import { areTokensEqual, getSwapRouterContractAddress } from 'utilities';
 
-import TEST_IDS from '../testIds';
+import SwapSummary from '../../SwapSummary';
 import { FormError } from '../useForm/types';
-import { useStyles } from './styles';
 
 export interface SubmitSectionProps {
   isFormValid: boolean;
@@ -34,7 +32,6 @@ export const SubmitSection: React.FC<SubmitSectionProps> = ({
   isSwapLoading,
 }) => {
   const { t } = useTranslation();
-  const styles = useStyles();
 
   const submitButtonLabel = useMemo(() => {
     if (isSwapLoading && Number(fromTokenAmountTokens) > 0) {
@@ -70,38 +67,6 @@ export const SubmitSection: React.FC<SubmitSectionProps> = ({
     return t('operationModal.repay.submitButtonLabel.repay');
   }, [isSwapLoading, fromTokenAmountTokens, isFormValid, formError, isFormSubmitting]);
 
-  const swapSummary = useMemo(() => {
-    if (!swap) {
-      return undefined;
-    }
-
-    const fromTokenAmountWei =
-      swap.direction === 'exactAmountIn'
-        ? swap.fromTokenAmountSoldWei
-        : swap.expectedFromTokenAmountSoldWei;
-    const toTokenAmountWei =
-      swap.direction === 'exactAmountIn'
-        ? swap.expectedToTokenAmountReceivedWei
-        : swap.toTokenAmountReceivedWei;
-
-    const readableFromTokenAmount = convertWeiToTokens({
-      valueWei: fromTokenAmountWei,
-      token: swap.fromToken,
-      returnInReadableFormat: true,
-    });
-
-    const readableToTokenAmount = convertWeiToTokens({
-      valueWei: toTokenAmountWei,
-      token: swap.toToken,
-      returnInReadableFormat: true,
-    });
-
-    return t('operationModal.repay.swapSummary', {
-      toTokenAmount: readableToTokenAmount,
-      fromTokenAmount: readableFromTokenAmount,
-    });
-  }, [swap]);
-
   return (
     <ApproveTokenSteps
       token={fromToken}
@@ -122,16 +87,7 @@ export const SubmitSection: React.FC<SubmitSectionProps> = ({
             {submitButtonLabel}
           </PrimaryButton>
 
-          {swapSummary && (
-            <Typography
-              data-testid={TEST_IDS.swapSummary}
-              css={styles.swapSummary}
-              variant="small2"
-              component="div"
-            >
-              {swapSummary}
-            </Typography>
-          )}
+          <SwapSummary swap={swap} type="repay" />
         </>
       )}
     </ApproveTokenSteps>
