+++ src/hooks/useOperationModal/Modal/RepayForm/__tests__/indexIntegratedSwap.spec.tsx
@@ -21,6 +21,7 @@ import en from 'translation/translations/en.json';
 import originalIsFeatureEnabledMock from 'utilities/__mocks__/isFeatureEnabled';
 
 import Repay, { PRESET_PERCENTAGES } from '..';
+import SWAP_SUMMARY_TEST_IDS from '../../SwapSummary/testIds';
 import TEST_IDS from '../testIds';
 import { fakeAsset, fakePool } from './fakeData';
 
@@ -364,7 +365,7 @@ describe('hooks/useBorrowRepayModal/Repay - Feature flag enabled: integratedSwap
 
     await waitFor(() => getByTestId(TEST_IDS.swapDetails));
     expect(getByTestId(TEST_IDS.swapDetails)).toMatchSnapshot();
-    expect(getByTestId(TEST_IDS.swapSummary)).toMatchSnapshot();
+    expect(getByTestId(SWAP_SUMMARY_TEST_IDS.swapSummary)).toMatchSnapshot();
   });
 
   it('updates input value to 0 when pressing on max button if wallet balance is 0', async () => {
