+++ src/hooks/useOperationModal/Modal/RepayForm/testIds.ts
@@ -2,7 +2,6 @@ const TEST_IDS = {
   tokenTextField: 'repay-modal-token-text-field',
   selectTokenTextField: 'repay-modal-select-token-text-field',
   swapDetails: 'swap-details',
-  swapSummary: 'swap-summary',
 };
 
 export default TEST_IDS;
