+++ src/hooks/useOperationModal/Modal/SwapSummary/testIds.ts
@@ -1,0 +1,5 @@
+const TEST_IDS = {
+  swapSummary: 'swap-summary',
+};
+
+export default TEST_IDS;
